Forms button completed
- Add the rest of the forms
